Handle empty bodies.
In some cases a code body may have only control-flow modifying statements such as 'break' or 'raise' but no statements that result in anything being added to a basic block.  This can be tricky as the effect of those control-flow statements must be attributed to blocks processed by an earlier Scope rather than blocks just created by the current Scope.

This change tries to deal with this situation by doing away with the notion of a single root block for a Scope which is linked to the rest of the graph after the Scope has finished building its portion of the tree.  Instead, the Scope can specify multiple roots which later code links to and each Scope is passed an earlier Scope's 'trajectory' which is the set of exit blocks which empty body control-flow statements will appear to originate from.