Improve exception strategy when using visitors.
The visitor pattern is incompatible with Java's checked exceptions.  Unfortunately, the AST visitor deals with this by making every method throw the Exception superclass.  This was infecting all the rest of the code.  Now we wrap calls to the visitor with try/catch blocks and only allow through exception type that the visitor implementations actually throw.  Anything else is unexpected and converted to a RuntimeException.  This isn't great (it still doesn't do exception checking for visitors) but as it didn't before, we lose nothing.